Refactor searchbar
